chore: add mutation test
