+++ CRA-demo/package.json
@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": false,
   "dependencies": {
-    "@apollo/client": "3.7.2",
+    "@apollo/client": "3.7.3",
     "graphql": "16.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-router-dom": "6.4.4",
+    "react-router-dom": "6.6.1",
     "react-scripts": "5.0.1",
     "web-vitals": "3.1.0"
   },
   "devDependencies": {
-    "@testing-library/cypress": "8.0.7",
-    "cypress": "11.2.0",
+    "@testing-library/cypress": "9.0.0",
+    "cypress": "12.3.0",
     "eslint-plugin-cypress": "2.12.1",
-    "start-server-and-test": "1.15.0"
+    "start-server-and-test": "1.15.2"
   },
   "scripts": {
     "cy:run": "cypress run",
