+++ CRA-demo/src/index.js
@@ -20,9 +20,7 @@ import ErrorInDeferredFragment from "./pages/errorInDeferredFragment";
 import ErrorNonNullableInDeferredFragment from "./pages/errorNonNullableInDeferredFragment";
 import ErrorNonNullableOutsideDeferredFragment from "./pages/errorNonNullableOutsideDeferredFragment";
 import ErrorTopLevelQueryField from "./pages/errorTopLevelQueryField";
-
-// mutation testing
-import PurchasePage from "./pages/purchase";
+import DeferredMutation from "./pages/deferredMutation";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -78,10 +76,7 @@ root.render(
             path="/error-top-level-query-field"
             element={<ErrorTopLevelQueryField />}
           />
-          <Route
-            path="/purchase"
-            element={<PurchasePage />}
-          />
+          <Route path="/deferred-mutation" element={<DeferredMutation />} />
         </Route>
       </Routes>
     </Router>
