+++ scripts/docker-compose.router.yml
@@ -2,7 +2,7 @@ version: "3"
 services:
   apollo-router:
     container_name: apollo-router
-    image: ghcr.io/apollographql/router:v1.5.0
+    image: ghcr.io/apollographql/router:v1.7.0
     volumes:
       - ./supergraph.graphql:/dist/schema/supergraph.graphql
       - ./router.yaml:/dist/config/router.yaml
